this and previous alg broken
test A through the document by path spits out the correct string. but using it as the string in the Puzzle constructor might be the problem. Thought a 3x3 was 51x51 and the max ship size was 49. I have no idea where those numbers came from as the string contained no number over 3.