tested and decided peer's ascii function suggestion
ascii - int conversion functions exist, but would require importing and calling a new function, and possible steps to extract just the single number to send along. All those steps, memory calls, and the existing solution is a simple subtraction. this is such a trivial case that there is no need for a function call and all that