Saved Early Performant Version Code
The earliest version of this project was just to be used as a performant coding framework. But, to use it more effectively as a code sample framework, I need to do some rewrites as optimization isn't as important as readability and extensibility. So the next version will have more standard function use and better data separation.