redesigned method flow
packageBoard is now a constructor helper method to build the data structure for the board and store it as a member of the object. there will be an overload for printBoard that will assemble a string version of the board to be used as necessary for debugging etc. After the overloads are complete, and testing is run successfully: the old methods and tests will be removed and replaced as necessary